requirements.txt
asgiref==3.5.0
auxlib==0.0.43
backports.zoneinfo==0.2.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.6
cloudinary==1.28.0
cmake==3.27.1
colorama==0.4.6
conda==4.8.3
conda-package-handling==1.7.0
cryptography==2.9.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0
django-cloudinary-storage==0.3.0
django-collectionfield==0.0.4
django-countries==7.2.1
django-decouple==2.1
django-filebrowser==4.0.1
django-filebrowser-no-grappelli==3.8.0
django-filebrowser-no-grappelli2==3.8.5
django-grappelli==3.0.3
django-import-export==3.2.0
django-jazzmin==2.6.0
django-parler==2.3
django-rosetta==0.9.8
django-taggit==2.1.0
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
djangorestframework==3.13.1
dlib==19.19.0
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gitignore==0.0.8
gTTS==2.3.2
gunicorn==20.1.0
idna==2.9
install==1.3.5
jsmin==3.0.0
MarkupPy==1.14
menuinst==1.4.16
MouseInfo==0.1.3
mss==9.0.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.4
odfpy==1.4.1
opencv-python==4.8.0.74
openpyxl==3.1.2
panda==0.3.1
Pillow==9.2.0
playsound==1.2.2
polib==1.1.1
protobuf==3.20.3
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycosat==0.6.3
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-gettext==4.0
python-magic==0.4.25
pytweening==1.0.7
pytz==2021.3
pywin32==227
PyYAML==6.0
requests==2.23.0
ruamel-yaml==0.15.87
six==1.14.0
SpeechRecognition==3.10.0
sqlparse==0.4.2
tablib==3.5.0
tokenize-rt==4.2.1
tqdm==4.46.0
tzdata==2021.5
Unidecode==1.3.6
urllib3==1.25.8
whitenoise==5.3.0
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
